refine how this component updates
debugging

add key prop to markup!

debugging

cleanup

cleanup

cleanup

cleanup
